% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plate_data8.R
\name{check_plate_data}
\alias{check_plate_data}
\title{Track and Process ELISA Plate Sample Data}
\usage{
check_plate_data(
  formatted_data,
  sample_ids,
  positivity_cutoff = 4.7,
  qc_percent = 10
)
}
\arguments{
\item{formatted_data}{Dataframe. Cleaned ELISA data from `format_elisa_data()`.}

\item{sample_ids}{Dataframe. Sample ID dataframe with sample_number and sample_id.}

\item{positivity_cutoff}{Numeric. Default is 4.7.}

\item{qc_percent}{Numeric. Percent buffer above/below cutoff to flag QC drift. Default is 10 (%).}
}
\value{
A dataframe with replicate values, means, CV, flags, and sample IDs.
}
\description{
This function processes cleaned ELISA data to calculate replicate means,
CVs, positivity flags, and quality control (QC) drift near the cutoff.
}
