% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_row_structure_function.R
\name{detect_row_structure}
\alias{detect_row_structure}
\title{detect row structure function}
\usage{
detect_row_structure(file_path)
}
\description{
This function searches the files (file_path) to see if sample ids were not included/missing per plate
it calculates the distance between the 'Layout' row and the '450' row
   if the distance is too close together, we know there are no sample ids
   and it gives a value of 2, instead of 3 (we expect three data-rows per plate-row)
}
\details{
'
}
