% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_files_function4.R
\name{extract_files}
\alias{extract_files}
\title{extract data from multiple files (w/ progress bar)}
\usage{
extract_files(file_list, error_summary)
}
\arguments{
\item{file_list}{Character vector. A list of file paths to be processed.}

\item{error_summary}{Dataframe. The output from `check_errors(file_list)`, containing files with errors.}
}
\value{
None. The extracted tables and sample IDs are assigned as separate dataframes.
}
\description{
This function extracts tables and sample IDs from valid files and automatically
assigns them to named variables in the global environment.
}
\examples{
file_list <- list.files(path = "data/", full.names = TRUE)
error_summary <- check_errors(file_list)
extract_files(file_list, error_summary)

}
