% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample_ids_helper_function.R
\name{extract_sample_ids}
\alias{extract_sample_ids}
\title{extract sample ids function}
\usage{
extract_sample_ids(file_path)
}
\description{
This function extracts the sample ids from the 'layout' /plate map table
}
\details{
this code will need the  keyword for this extraction function to work
   'Layout' for the platemap table
   if layout is missing (even if the sample ids are present it will mark them as NA)

 this key word should be in the first column in the excel file,
 the sample ids need to be laid out a certain way for this function to work
   an example output can be found at this link: [LINK]

 the 25x13 data table (24x12 not counting column/row headers) should start:
       3 rows down and 1 column right of the layout keyword


 we will call this function in the same for loop as our extract tables function

 this will output each plate as a dataframe with 8 rows
 we will call the file_list and pull each dataframe per plate into one large list in a for loop
 then we'll combine all the individual dataframes into one list after the for loop
}
