% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables_helper_function.R
\name{extract_tables}
\alias{extract_tables}
\title{extract tables function}
\usage{
extract_tables(file)
}
\description{
This function extracts the tables (450, blank 450, concentration) from the excel file output
We have a separate function that selects sample ids from the plate map table
}
\details{
you need to manually set your folder_path (where your files are) and file_type (.xlsx)
make sure you close all the excel files from the selected folder path!

make sure your files are named numerically and the first underscore is after the numbers,
 in this case:
          'Plate 001_date_techinitials.xlsx'
          'Plate 002_date_techinitials.xlsx'


folderpath <- "C:/Users/tti2/OneDrive - CDC/Data Cleaning/ELISA Data Management/uganda"
file_list <- list.files(path = folderpath, pattern = "*.xlsx", full.names = TRUE)

each file will need three keywords for this extraction function to work
   '450' - for the 450 table
   'Blank 450' - for the BLank 450 table, which we might not actually need?
   '[Concentration]' - for the conc table, which we definitely need

 these three key words should all be in the first column in the excel file,
 an example output can be found at this link: [LINK]

 the 8x12 data table (9x13 counting column/row headers) should start:
       3 rows down and 1 column right of each keyword
       in the case of the 450 table, the 'Actual Temperature'
        can be present or absent, up to user


 this will output each plate as a dataframe with 24 rows
 we will call the file_list and pull each dataframe per plate into one large list in a for loop
 then we'll combine all the individual dataframes into one list after the for loop
}
