% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_elisa_data6.R
\name{format_elisa_data}
\alias{format_elisa_data}
\title{Format and Clean ELISA Data}
\usage{
format_elisa_data(all_data_combined)
}
\arguments{
\item{all_data_combined}{Dataframe. The raw ELISA data.}
}
\value{
A cleaned dataframe with numeric values properly formatted.
}
\description{
This function renames columns, cleans numeric values, and standardizes ELISA plate data.
}
\examples{
formatted_data <- format_elisa_data(all_data_combined)
head(formatted_data)

}
