% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sample_ids_function5.R
\name{format_sample_ids}
\alias{format_sample_ids}
\title{Format Sample IDs into a Structured Dataframe}
\usage{
format_sample_ids(all_sample_ids_combined)
}
\arguments{
\item{all_sample_ids_combined}{Dataframe. A dataframe containing extracted sample IDs.}
}
\value{
A formatted dataframe with paired replicates and structured sample numbers.
        If duplicate sample IDs are found, a warning message is displayed.
}
\description{
This function processes raw sample ID data, extracts plate numbers, pairs replicates,
removes control samples, assigns sequential sample numbers, and checks for duplicate sample IDs.
}
\examples{
formatted_sample_ids <- format_sample_ids(all_sample_ids_combined)
head(formatted_sample_ids)

}
