% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_list_function1.R
\name{get_files}
\alias{get_files}
\title{select folder and file path for samples}
\usage{
get_files(folderpath, file_type = "xlsx")
}
\description{
#specify a folder path
folderpath <- "path/to/folder"
}
\details{
attempt to maintain one filetype

#identify your folder containing elisa excel files
file_path <- get_files_by_type(folderpath, file_type = "xlsx")

#dentify your folder containing elisa csv if they're not excel files
file_path <- get_files_by_type(folderpath, file_type = "csv")
}
