% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files_function.R
\name{process_elisa_files}
\alias{process_elisa_files}
\title{process samples function}
\usage{
process_elisa_files(file_list)
}
\description{
This loop processes each plate using our two previous functions
and outputs our data into two dataframes (sample ids and plate data) and one list (error log)
process_elisa_files(file_list)
}
